Ürünler alınıyor.

+++ Products.java
@@ -2,9 +2,9 @@ public class Products {
     private String name;
     private double price;
     private int stock;
-    private float rating;
+    private double rating;
 
-    public Products(String name, double price, int stock, float rating) {
+    public Products(String name, double price, int stock, double rating) {
         this.name = name;
         this.price = price;
         this.stock = stock;
